requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via
    #   django
    #   django-minify-html
chardet==5.2.0
    # via reportlab
confusable-homoglyphs==3.2.0
    # via django-registration
crispy-bulma==0.11.0
    # via -r requirements.in
django==5.0.7
    # via
    #   -r requirements.in
    #   crispy-bulma
    #   django-appconf
    #   django-crispy-forms
    #   django-markdownx
    #   django-minify-html
    #   django-registration
    #   django-select2
    #   django-username-email
django-appconf==1.0.4
    # via
    #   django-imagekit
    #   django-select2
django-capture-tag==1.0
    # via -r requirements.in
django-crispy-forms==2.2
    # via
    #   -r requirements.in
    #   crispy-bulma
django-imagekit==5.0.0
    # via -r requirements.in
django-markdown-deux==1.0.6
    # via -r requirements.in
django-markdownx==4.0.7
    # via -r requirements.in
django-minify-html==1.8.0
    # via -r requirements.in
django-registration==3.4
    # via -r requirements.in
django-select2==8.2.0
    # via -r requirements.in
django-username-email==2.5.4
    # via -r requirements.in
django-webpack-loader==3.1.0
    # via -r requirements.in
dsnparse==0.2.1
    # via -r requirements.in
gunicorn==22.0.0
    # via -r requirements.in
markdown==3.6
    # via
    #   -r requirements.in
    #   django-markdownx
markdown2==2.3.9
    # via django-markdown-deux
minify-html==0.15.0
    # via django-minify-html
packaging==24.1
    # via gunicorn
pdfrw==0.4
    # via -r requirements.in
pilkit==3.0
    # via
    #   -r requirements.in
    #   django-imagekit
pillow==10.4.0
    # via
    #   -r requirements.in
    #   django-markdownx
    #   pilkit
    #   reportlab
psycopg2-binary==2.9.9
    # via -r requirements.in
pydash==8.0.1
    # via -r requirements.in
pypdf==4.3.0
    # via -r requirements.in
reportlab==4.2.2
    # via -r requirements.in
sqlparse==0.3.1
    # via django
typing-extensions==4.12.2
    # via pydash
